+++ src/parser/DxfParser.js
@@ -4,7 +4,7 @@ import AUTO_CAD_COLOR_INDEX from './AutoCadColorIndex';
 import Face from './entities/3dface';
 import Arc from './entities/arc';
 import AttDef from './entities/attdef';
-import AttRib from './entities/attrib'
+import Attribute from './entities/attribute'
 import Circle from './entities/circle';
 import Dimension from './entities/dimension';
 import Ellipse from './entities/ellipse';
@@ -34,7 +34,7 @@ function registerDefaultEntityHandlers(dxfParser) {
     dxfParser.registerEntityHandler(Face);
     dxfParser.registerEntityHandler(Arc);
     dxfParser.registerEntityHandler(AttDef);
-    dxfParser.registerEntityHandler(AttRib);
+    dxfParser.registerEntityHandler(Attribute);
     dxfParser.registerEntityHandler(Circle);
     dxfParser.registerEntityHandler(Dimension);
     dxfParser.registerEntityHandler(Ellipse);
