+++ src/parser/entities/attribute.js
